knowledge-graph/src/backend/pyproject.toml
[tool.poetry]
name = "knowledge-base-backend"
version = "0.1.0"
description = "Enterprise Knowledge Base Management System Backend"
authors = ["Knowledge Base Team <team@hktlora.com>"]
readme = "README.md"
packages = [{include = "knowledge_base_backend", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^3.0.0"}
flask-restx = "^1.3.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-jwt-extended = "^4.6.0"
flask-cors = "^4.0.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.4"
elasticsearch = "^8.11.0"
jieba = "^0.42.1"
nebula3-python = "^3.8.0"
networkx = "^3.2.1"
igraph = "^0.11.3"
minio = "^7.2.0"
pydantic = "^2.5.0"
marshmallow = "^3.20.1"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
python-multipart = "^0.0.6"
pypdf2 = "^3.0.1"
python-docx = "^1.1.0"
openpyxl = "^3.1.2"
pyyaml = "^6.0.3"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-flask = "^1.3.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
requests = "^2.31.0"
selenium = "^4.15.2"
webdriver-manager = "^4.0.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "system: System tests",
    "slow: Slow running tests",
    "security: Security tests"
]